* ext/dl: revert dl with libffi because it can't run on mswin now.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@26764 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/dl/cptr.c
@@ -104,7 +104,7 @@ rb_dlptr2cptr(VALUE val)
     else{
 	rb_raise(rb_eTypeError, "DL::PtrData was expected");
     }
-
+    
     return ptr;
 }
 
@@ -186,7 +186,7 @@ static VALUE
 rb_dlptr_s_malloc(int argc, VALUE argv[], VALUE klass)
 {
     VALUE size, sym, obj;
-    long s;
+    int   s;
     freefunc_t f;
 
     switch (rb_scan_args(argc, argv, "11", &size, &sym)) {
